requirements.txt
pandas==2.2.0
jupyter==1.0.0
tqdm==4.66.2
pinecone-client==3.0.2
langchain==0.1.6
python-dotenv==1.0.1
openai==1.12.0
tiktoken
langchain-openai==0.0.5
sentence-transformers==2.3.1
gradio==4.18.0